requirements.txt
base58==2.1.0
cytoolz==0.11.0
ecdsa==0.16.1
eth-hash==0.3.1
eth-typing==2.2.2
eth-utils==1.10.0
pycryptodome==3.10.1
rlp==2.0.1
six==1.16.0
toolz==0.11.1
